requirements.txt
Django==1.8.5
djangorestframework==3.2
Markdown==2.6.2
django-filter==0.11.0
django-guardian==1.3
Pillow==2.9.0
gunicorn==19.3.0
psycopg2==2.6.1
django-environ==0.4.0
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
django-allauth==0.23.0
